Siof/lh utf8 fix port (#42)
Resolve utf8 conversion buffer overflow when characters are 4 octets wide

Worth noting that the main issue was that utf8::distance does not return the size required to store all the components of the UTF string - only the number of symbols. You cannot know the size required until you actually do the conversion.

Therefore, since we calculated a 48 byte string full of 4 octet symbols (4 bytes per symbol in UTF-8) to only require 12 symbols (dist == 12), we only allocated 12 bytes for the wstring. In reality, it takes 48 / 4 * 2 = 24 bytes in the wstring to store 12 two-byte wide symbols. The same issue exists for any symbols that are 3 bytes wide in UTF-8.

Hence the solution is to allocate the wstring the same size as the input string, and resize appropriately once widening has taken place.